Use emptyDir in order to work around leak
With GitLib2 we have the issue that the storage slowly, but eventually runs full.
Using emptyDir works around this issue,
as it restarts the pod and creates a new volume when the original volume is full.
Plus, we never needed our volume to be persistant anyway.
All critical data is saved in the manifest repository.