Temporarily going back to use a persitant volume
Currently we still have a cache, and that cache only works when the data is stored beyond pod restarts.

In the future we will remove the cache and go back to emptydir, but for now, we need the cache working.

This is almost an exact revert of commit fc057869dc4656cc59e5db195f1cb4879c720add.