Add Environment Groups to getOverview response (#462)
This adds a new field `EnvironmentGroups` to the server response.
The old field `Environments` still exists, but is deprecated. Logically, they contain the same information, but EnvironmentGroups is structured better for use in the UI.

This commits moves over functions like `calculateEnvironmentPriorities` to the server, that where implemented on client side before.