Bugfix: Do not require authentication in frontend-service health check (#467)
Before this change, when `AzureAuth` was enabled, the frontend services's health check did not work, it only responds with `Invalid authorization header provided`.
With this change, the health endpoint does not requires authentication anymore. It does not matter anymore if `AzureAuth` is enabled or not.

Version 0.4.52 should not be used on Azure.