Fix sorting of EnvironmentGroups
This change has no real effect on real users, because the edge case behavior should never happen in real life. The required configuration is nonsensical for the real world.

Our algorithm to calculate distance to upstream depends on the order of the given EnvironmentGroups.
To be sure that the result is always the same, we have to iterate over environmentGroups in a sorted manner.
However, the `range` function in golang does not guarantee an order on maps.
This was only detected now, because it happens only sometimes, and it only affects an edge case were environments are configured with a circular dependency.
Only the test "Two Environments with a loop" could reproduce the error (sometimes).

Also the resulting priority in this edge case was changed to "666", to make it a little more obvious that something is wrong.