Render environment groups in UI (#465)
This renders environment groups in the UI.

Env labels have been changed to display groups. They have 2 numbers:

The number of envs that have the same release deployed (within the env group and for a given service)
The total number of environments in the env group.
Env labels are displayed as groups in:

- The Home page (ReleaseCard)
- The history page (ReleaseCardMini)
But not in the ReleaseDialog. Here we need to display all envs always.
However, this change buckets the environments in one group closer together, giving them a bigger margin.