Refactoring: Require return types in new ui
Added rule '@typescript-eslint/explicit-function-return-type' which requires all functions to have an explicit return type.

Also added a specific .eslintrc for the old ui, which disables this rule, because we do not want to spend much time with the old ui anymore.

Adapted all functions that didn't have a return type.

This is a refactoring, it does not change any behavior.

SRX-2DXE9T