Added warning if there are no envs configured during startup. (#502)
Kuberpult cannot work properly without at least one environments configured, so we print a warning on startup, if none are found.

Usually this is an error (at least when running locally). However, sometimes it is intended to start up without environments (and then add them dynamically via the "createEnvironment" endpoint).
Therefore, it's just a warning.