Fix undeploy for apps that are not in all environments (#555)
This fixes a bug in the "undeploy" request.
The issue was that kuberpult checked if the app was deployed (with version "undeploy") in all environments. This was a prerequisite to run "undeploy".

However, some apps are not supposed to be deployed in all environments. (It's an edge case, but it can happen, especially when environments are added later on).

Before:
 * The "undeploy" version of the given app must be deployed in all environments.
After:
 * If the app exists in an env, it has to be the "undeploy" version.

This is a fix for https://github.com/freiheit-com/kuberpult/issues/451

SRX-M2E91H