Fix bug in app deletion when there are app locks
When there is a lock on the app, and gets a request to delete an app, the server returns an error.

Now the client gets the information (in advance as part of the undeploySummary) and will not show the button if there is an app lock.

Env locks are not considered for deletion of apps.