Fix: Allow emptying manifest of individual environments (#650)
Before, when calling the release endpoint with an empty manifest, kuberpult would not write the manifest to the manifest-repository.
Now it writes a file with the a single space. This works around an issue in the WriteFile function.

We had a similar issue when using the "undeploy" function, which had already been fixed.