Separate tags for services
This is *technically* a breaking change.

This change allows to specify different docker image tags for the cd service and the frontend service.

However, this is only the case if you specify the "tag" in the helm chart.
It is rather unlikely that you do this.
In almost every case, kuberpult should run with the same version for all microservices (cd services and frontend service).