Refactored error handling
Now the batch endpoint can easily distinguish between internal and public errors.
Public errors are sent back to the client.
Internal errors are logged only and the client just receives "internal error"

To achieve this, the validation functions have all been inlined into the transformers.
This way the transformers are self sufficient, and validations do not need to be duplicated either.