Render warnings for unusual deployment situations
We now render 2 warnings: If "prod" is ahead of "staging", that is unusual.
Probably someone deployed a bugfix on prod, but forgot to lock staging.

Second warning: if one env has no version deployed but the downstream of it has one.
This means the next release train would get rid of it.