Also lock when deploying manually (#732)
Manual deployments happen usually when there is an important bugfix to be deployed. In that scenario, usually a lock is a good idea. With this change, the "deploy" button is replaced by "deploy & lock".
The user can still remove the lock from the "planned actions" if they really don't want it.