Make git author configurable in helm chart & refactor context usage
This is technically not a breaking change, but it is recommended to now set the author and email when invoking the /release endpoint. Note that both headers must be base-64 encoded. See create-release.sh for a valid example call with headers. If you do not provide these headers, the git author will be set to "empty".

New helm chart variables can be set, but they come with the same default values as before:

* git.author.name
* git.author.email

This also refactored how to access the context.

The cd-service now always* returns an error when the context should be there but is not, instead of always falling back to default values.

Exception is the release endpoint, as this would be a breaking change. (And of course /health does not require any headers)

The frontend-service still uses the headers if supplied, but falls back to the default values that are configured in the helm chart.

This effects all endpoints, http and grpc.

The context functions (formerly "Extract", etc) have been refactored, so they do exactly one thing, e.g. reading from one context.
There are 3 options: grpc context, go context and http headers. All have a function for reading and writing respectively.

Story SRX-6SMIU3