Replaced /release http endpoint in cd-service with grpc (#814)
Before, we had a release http endpoint in the cd-service and the frontend-service had it too, just redirecting to the cd-service.
Now we have a release http endpoint in the frontend-service, and it uses a new grpc endpoint to call the cd-service.

This is technically a breaking change, because you may have used the cd-service /release endpoint directly.
However, if you are using the helm chart, then this will not be breaking, because it exposes the right endpoint correctly.

Only effective change is that the auther parameters for the release endpoint are not optional.
If not provided, the ones in the helm chart will be used. However, this may change in the future, so we recommend to continue to provide the author parameters.

Also added some integration test, because this endpoint is now only testable with an integration test, as 2 services are involved. This required setting up a gpg keypair, because we want to test the signatures in /release as well.

Co-authored-by: Sven Urbanski <sven.urbanski+git@freiheit.com>