Allow ArgoCd to use webhooks
When having thousands of ArgoCd apps, it becomes problematic to sync the whole repository every n minutes (3 by default).
This PR introduces an annotation to tell argo: If there is a webhook incoming, compare the paths in the webhook with the paths in this annotation. ArgoCd only needs to synchronize the files that are in that path.
If not set, argoCd will synchronize every file in the whole manifest repo.