Tell ArgoCd to sync apps (#872)
With this feature, kuberpult sends webhooks that imitate github webhooks to argoCd, which trigger argoCds refreshs.

This is intended to reduce load on ArgoCd. Normally, argoCd renders all manifests again whenever there is a new commit in the manifest repository. However, with Webhooks, it only checks the actually changed files that are part of the Webhook body.

Note that this does not work for all cases, like deletion of environments from apps.

So you can make the "timeout.reconciliation" in argoCd bigger (e.g. 30min instead of the default 3min) - however you should not fully disable it. See https://argo-cd.readthedocs.io/en/stable/operator-manual/high_availability/#argocd-application-controller

To enable this kuberpult feature, set `argocd.server` in the helm chart to a hostname that reaches argoCds "argocd-server", usually this should be `http://argocd-server.${ARGO_NAMESPACE}.svc.cluster.local`.

If you use https and the certificate for the https connection is not accepted, set `argocd.insecure: true`.
This is usually safe, under the assumption that kuberpult and argoCd run on the same cluster anyway.

This feature is intended to be used when Webhooks are not an option for you (for example due to firewall policies). The usual way to solve this is configure a webhook in your git provider to communicate to argoCd. See argoCd docs on that here: https://argo-cd.readthedocs.io/en/stable/operator-manual/webhook/

Caveat: Kuberpult only retries failed connections a very limited number of times. That means, if argoCd does not react temporarily, webhooks may get lost, and therefore argoCd does not sync until the reconciliation timeout.