Submit user data via grpc message instead of header
Headers are genereally not safe with special characters.
Instead of encoding the header with base64, we move the data to a normal grpc message (ProcessBatch), because this is the only request where we have the userdata anyway.

This fixes the bug https://github.com/freiheit-com/kuberpult/issues/874 that happened on azure when users have non-ascii usernames - but it is very hard to verify this.