UI: Show warning before deploying
If the user wants to deploy an app that has a lock, or
if the user wants to deploy to an environment that has a lock,
we now show a warning, because generally locks should not be ignored.

The dialog shows which locks are the "offenders".

The user can still choose to ignore this, but now they have been warned.