Add environment group name to kuberpult events
We need the environment group name for emitting the metrics with the
correct key. The other method of the service could also supply this, but
it's actually more complicated because `GetVersion` can access historic
states and the environment group of an environment may have changed in
the meantime. Therefore, I only added this data to the event, which is
100% sufficient for us now.