Remove chart testing from CI
It is not working properly. It has its own logic to decide if it needs to run, and often it just doesnt.
I suspect this has not been working for a long time.
It is not a huge loss, because we have integration tests that also install our helm chart.