Bugfix: Allow selection of envs to delete from an app (#1098)
This fixes https://github.com/freiheit-com/kuberpult/issues/1091

We were relying data from a javascript event to get the index of the selected environment.
This meant that the onClick handler had to be added at the correct element in react (and in an earlier refactoring the wrong element was chosen).

It turns out, we do not need that at all. We can just pass along the id of the environment as a separate argument to the callback. The whole javascript event is not necessary anymore, so it was removed.

Co-authored-by: Sven Urbanski <sven.urbanski+git@freiheit.com>