Fail health check when too many git pushes fail
From real life experience, it seems there are some cases where a restart of the cd-service helped to fix a hanging git push processes.

This automates the restart by adding the number of failures to the health check. We restart when we see more than 3 git push errors.

Reference: SRX-976IRJ