Restart pod when a queued action takes longer than 3*networkTimeout.
This changes the cd-service so that it automatically exits when a queued action (like a release train), takes longer than 3 times the configured networkTimeout.

The problem is that sometimes "git push" actions take a long time. They cannot be run in parallel, and they can not properly be cancelled either.
So as a last resort, we just exit, and let kubernetes restart the pod.