feat: save commits info in manifest repo (#1294)
Here we add logic to `CreateApplicationVersion`, `UndeployApplication`,
and `CleanupOldApplicationVersions` transformers so that there is a
separate `commits` directory in the manifest repository that maintains
the currently-tracked commits.

The new path has the format `commits/<first two characters of the
hash>/<remaining letters of the hash>/applications/<app>/.empty`. It's
possible that there will be more than one application in one commit's
subtree. The `.empty` file is there so that the path is tracked by Git.

This is a prerequisite for implementing a commit info page in the UI
later. We make data about commits available directly, as opposed to
through the Overview service as it's inefficient.


---------

Co-authored-by: Ahmed Nour <125875466+ahmed-nour-fdc@users.noreply.github.com>
Co-authored-by: Sven Urbanski <sven.urbanski@freiheit.com>