feat: Group the rollout status in the release tooltip by env-group (#1313)
Listing the rollout status for each environment can make the tooltip to
big in projects with many environments. Instead, we group the rollout
status by environment group, which results in much fewer entries. The
rollout status is accumulated in the same way as the overall rollout
status displayed in the card, i.e. the most important rollout status is
chosen.

![image](https://github.com/freiheit-com/kuberpult/assets/50741717/9fdb09f4-2546-4f22-ad21-53bd2559fd6d)