fix: environment cards colors (#1321)
In this PR we change the rules for environment cards colors in the
environments page.

Environments have a
[`priority`](https://github.com/freiheit-com/kuberpult/blob/19d0f430f3948526694b099e4f9be295afc89754/pkg/api/v1/api.proto#L337)
property that indicates where they lie in a chain of upstream relations.
Currently the color of an environment card is
[determined](https://github.com/freiheit-com/kuberpult/blob/4a3eaa34cd55732df7b16da61503fe54cbb0de9f/services/frontend-service/src/ui/components/EnvironmentCard/EnvironmentCard.tsx#L40)
based on this property.

The color of environment groups is determined by the priority of the
first environment in that group, which is erroneous: environments in one
group have the same _distance to upstream_, but not necessarily the same
_priority_. This leads to situations of an environment group its
environments having different colors which looks awkward in the UI.

| ![](https://i.imgur.com/LGvQj6t.png) |
| :---: |
| _Figure 1: awkward UI colors, notice that both `testing-1` and
`testing-2` have different colors. This is because `testing-2` has not
downstream, so it is assigned the `PRODUCTION` priority level._|

In this PR we:
1. Introduce a new field to the environment group struct, called
`priority` of type `Priority`. Despite the similarity, this field for
environment groups is calculated in a different way from that of
environments such that the priority is determined based on the distance
to upstream and not the priorities of environments. The reason for
reusing the same name and the same type is to keep changes to the
frontend to a minimum.
2. Change the frontend logic for determining the colors of environment
and environment group cards to be based on the newly introduced
`EnvironmentGroup.priority` field.

Now it looks like this:

| ![](https://imgur.com/LUcbmqH.png) |
| :---: |
| _Figure 2: fixed colors in the UI_|

---------

Co-authored-by: Sven Urbanski <sven.urbanski@freiheit.com>
Co-authored-by: Leandro Salgado <110404621+leandro-freiheit@users.noreply.github.com>