chore: initialize all struct in rollout service (#1403)
This is a prerequisite PR for #1396 

Here we address or silence all the errors generated by exhaustruct in
the CD service. The decision for whether to address or silence an error
was the mere convenience; if a struct initialization seemed too large or
too complicated, the error was silenced, otherwise it was addressed.