feat: add `exhaustruct` to CI (#1396)
Here we add the
[`exhaustruct`](https://github.com/GaijinEntertainment/go-exhaustruct)
linter to CI. The CI will fail if there are uninitialized struct fields
in Go code (excluding test and generated files).

Note that `golangci-lint`, which is currently executed in the CI, does
[support](https://golangci-lint.run/usage/linters/#exhaustruct)
`exhaustruct`. However, there seems to be an issue when running
`exhaustruct` through `golangci-lint`: it doesn't respect the
`//exhaustruct:ignore` comments. Therefore, we will stick to running
`exhaustruct` separately.