fix: make colors between `EnvironmentCard` and `EnvironmentChip` consistent (#1432)
Currently environment colors are inconsistent between the environments
page and the release dialogue, as shown below.

| ![](https://i.imgur.com/LchXL1G.png) |
| :---: |
| _Figure 1: the environments page (left) and the release dialogue for
an arbitrary app (right). Note that `development` has different colors
in the two pages_|

These colors are determined by the `priority` field of `Environment` and
`EnvironmentGroup` types; the environments page refers to the
`EnvironmentGroup` type, while the release dialogue refers to the
`Environment` type. This inconsistency was introduced in #1321.

This PR fixes this by modifying the logic of calculating the CSS style
in the `EnvironmentChip` component (which consequently determines the
CSS style for the color) to be similar to that of the `EnvironmentCard`
component by referring to the `priority` field of the group of the
environment rather than the environment itself.

Now it looks like this:

| ![](https://i.imgur.com/lj59pGu.png)|
| :---: |
| _Figure 2: the colors now_ |