fix: add update check for update requests (#1423)
Currently we are sending requests to the argo CLI when there are apps to
be created or to be updated. The problem is that even if an application
is not changed, we are still sending an update request. We should not
send any request if the application does not change.