fix: reduce likelihood of gpg crashing (#1443)
We had some varying issues with integration test in earthly.

1) Use ubuntu as base image:
Quite often we had an issue that the gpg export command complains that a
gpg lockfile exists. Which should never happen, because we always run
one gpg command at a time. We didn't find the root cause, but this issue
seems to not happen at all when using ubuntu as a base image.

2) Change the order of the RUN commands in earthfile: 
When working locally, it can happen that files like
charts/kuberpult/kuberpult-keyring.gpg exists.
In the past, the earthfile would take those, due to the order of the
COPY commands.
We now write the kuberpult key ring into /kp instead of the local
directory to avoid this issue.