feat: limit number of batched commits per push (#1427)
Theoretically we could end up in a situation where we are trying to fit
an infinite amount of commits in one push. That is probably not an ideal
thing to do. There should be a limit on this, especially in environments
with high concurrency.

This change adds a new config option `.git.maximumCommitsPerPush` to the
helm chart that is used as an upper limit for the number of commits per
push.