fix: refactor tests to be more specific (#1487)
the test had both an env lock and an app lock, so it wasn't that clear
which was necessary. The test is now split up into 2 tests: one with
only an app lock and one with only an env lock.
Still, even more tests are needed, specifically for
https://github.com/freiheit-com/kuberpult/issues/1486

Also slightly formatted an error message.