feat: release train prognosis (#1476)
In this PR we add an endpoint for _release train prognosis_, which is
essentially a plan of what applications will be deployed where and what
application will be skipped (and the reason for skipping them) _in case_
the release train was to be triggered.

It is possible the prognosis returns an error, for example if the
environment configuration was not found. Note that, however, it is
_still_ possible for the release train to fail even if the prognosis
reports a certain plan. That would be the case in case the application
deployment returned an error.

The prognosis is available on the
`/api/environments/<environment_name>/releasetrain/prognosis` endpoint
(only `GET` methods are accepted). Example:

```
$curl http://localhost:8081/api/environments/devgroup/releasetrain/prognosis
{"development":{"Outcome":{"SkipCause":4}},"development2":{"Outcome":{"AppsPrognoses":{"prognoses":{"google":{"Outcome":{"SkipCause":1}},"photoshop":{"Outcome":{"SkipCause":1}},"windows":{"Outcome":{"SkipCause":1}}}}}}}
```

The response you see above is the serialization of the protobuf response
message. In the example above, `development` environment is being
skipped due to _cause number 4_, which according to the protobuf enum,
says that environment is locked. the environment `development2` will not
be skipped. However, all applications in that end will not experience a
deployment since there are no new applications upstream (upstream is
`latest`).

On the other hand,

```
$curl http://localhost:8081/api/environments/staging/releasetrain/prognosis
{"staging":{"Outcome":{"AppsPrognoses":{"prognoses":{"google":{"Outcome":{"DeployedVersion":13}},"photoshop":{"Outcome":{"SkipCause":1}},"windows":{"Outcome":{"SkipCause":1}}}}}}}
```

This one says that version 13 of `google` will be deployed on `staging`.

---------

Co-authored-by: Sven Urbanski <sven.urbanski@freiheit.com>