feat: add primitive CLI client for the `/release` endpoint (#1504)
This PR introduces the Kuberpult Client, a command line client and
library for interacting with Kuberpult.

This PR only implements a very simplistic version of the `release`
action of Kuberpult, it just accepts an application name, environment
names, and manifests for these environments, and calls the `/release`
endpoint of Kuberpult with these parameters, like the following

```
kuberpult-client release --application myapplication --environment dev1 --manifest dev1-manifest.yaml --environment dev2 --manifest dev2-manifest.yaml
```

In the current state the client only interacts with Kuberpult on
`localhost`.

The CI consists of a simple test, build, and push actions, written from
scratch. The executable is also shipped as a Docker container
`kuberpult-client`.