feat: write every transformer request to the DB
This adds a new "ESL" (event sourcing light) table. This tables stores the input data of every transformer (which is basically every grpc call) in a table.
This is handled outside the transformer. Additionally this database operation is part of a transaction. The transaction is passed to every transformer (which is why this PR is big).

Usually we group parallel requests, so that we need to `git push` only once - however, that does not work with the concept of "committing the transaction once the `git push` succeeeded.
Therefore this PR also disables the queuing for requests, if the DB is enabled.

Ref SRX-D1PBPL