fix: introduce generated sequence for esl table
This is necessary because otherwise there is not total order on the rows of the esl table.
Unfortunately, it is note possible to write an SQL statement that works for both sqlite and postgres.
Therefore, we split the migrations into sqlite and postgres and work with symlinks for cases where the SQL statements are the same - which is most of the time