fix: do not reset application sync policy on update (#1612)
When we do an update using the rollout-service, we always enable the
auto-sync. If there is an application that has its auto-sync disabled
manually for hotfix purposes, kuberpult will override this almost
instantaneously on the next consume of the overview. For this purpose,
we make the diff between applications on the update call more granular,
excluding the sync policy.

---------

Co-authored-by: Leandro Salgado <leandro.salgado_ext@mail.schwarz>