feat: scan for secrets (#1631)
Here we add a GitHub workflow that checks for secrets in the repository.

The workflow runs [Trivy](https://trivy.dev/) in filesystem (`fs`) mode
with the `secrets` scanner. We further added some configuration to
disable all builtin allow rules.

In case an alert was raised by Trivy, a Slack message is sent to us via
a webhook.