feat: add release train prognoses to the UI (#1647)
This PR adds a UI page for displaying the release train prognosis on a
certain environment at `/ui/environments/<env>/releaseTrain`. It is
accessible from a link on the environments group called "Release train
details", as shown.

|
![image](https://github.com/freiheit-com/kuberpult/assets/156064055/65650076-c031-4b0d-8af7-08959477c54c)|
| :---: |
| _figure 1: the environments page with the new link to the release
train prognosis page_|

The release train prognosis page contains a table with an entry for
every application stating what happens to the application in case a
release train triggered on this environment. It looks like this.

|
![image](https://github.com/freiheit-com/kuberpult/assets/156064055/efc14acf-1dfd-450b-85a1-a24ae5b3f745)
|
| :---: |
| _figure 2: a sample of the release train prognosis page_ |

There is a special case when the release train on an environment is
skipped (for example, because it is locked). In which case no table is
rendered and a descriptive message is shown instead.

|
![image](https://github.com/freiheit-com/kuberpult/assets/156064055/280bc649-a57a-4701-b4e4-eca0fdfbe19c)
|
| :---: |
| _figure 3: release train prognosis page when an environment is locked_
|