feat: use priority from envGroup for dora events, include canary envs (#1727)
The `priority` field on env level is discouraged and not used in the
frontend anymore. However, it's still used in the rollout service for
deciding whether to send dora events. Also, we discussed whether
"production" in the dora sense should include canary environments and
came to the conclusion that they should indeed be included.