feat: Add full name override for DEX resources (#1734)
When installing multiple instances of Kuberpult using Dex on the same
cluster, there are namespace independent resources such as ClusterRoles
that need to have a different naming.

Dex allows this by setting the fullNameOverride that allows to provide a
custom naming for all Dex resources. Kuberpult needs to make the
DexServiceURL configurable and do not use a constant value.