fix(db): replace sequence prediction (#1764)
The original attempt was to read "last_value" of a squence to determine
the *next* Primary Key and also write it into the json blob.
    
This was error prone, and especially failed on the first startup (with
an "empty" sequence).
    
This PR changes this approach. It was necessary to add an interface
function to set the transformer ID.
    
Background: The transformer ID is required for the commit events, so
that we can easily find all commit events of one transformer.