feat(cd-service): Cache overview response (#1751)
A new table called `overview_cache` is created. It has a `json` column
which contains the result of the overview service.
Whenever the overview service is called it will check the
`overview_cache` table. It gets the latests row's json and sends it as a
response. If there isn't any rows in the table, it will create one.
After any change in the database, the `overview_cache` will be updated.