fix: fix-lock-prevented deployment & overview error (#1798)
When creating an undeploy application version and an lock existed, a
lock prevented deployment event was generated. Undeploy versions have no
commit id and events require one. Fix was to not generate this event (by
throwing an error)if we can't find a commit ID (equal to what we did
with the manifest repo)

Also, the overview wasn't getting correctly updated.

REF: SRX-0E8EKB