feat: Allow team rbac perm (#1822)
# BREAKING CHANGE
Now to do any action related to an application and hence a team the user
must
be present on the team.csv file. The default behavior for when the file
is empty
is for no one to have permissions.  
## Instructions 
For the user with email "test@mail.com" to have permissions for the team
"test-team" this should be included in the team.csv:
"test-team,test@mail.com"
PR: https://github.com/freiheit-com/kuberpult/pull/1799

---------

Co-authored-by: João Correia <joao.correia@freiheit.com>