feat: optimize db requests (#1831)
Reduced number of DBSelectDeployments calls done by the
findOldApplicationVersions function. Created a new table that, for each
app, stores the releaseVersion deployed on each environment, reducing
the number of requests to the database by X, where X is the number of
environments.
Reduced the number of DBSelectEnvironment calls done by the
checkUserPermissions function. Instead of requesting all environment
configs, looping through them and finding the other we are looking for,
we simply request the one we are looking for, reducing the number of
requests to the database by X fold, where X is the number of
environments.

Ref: SRX-XT0P80