fix(db): write to cutoff table if writeEslTableOnly=true
If we were previously running with `db.writeEslTableOnly=true`, but now are running with
`db.writeEslTableOnly=false` (which is the recommended way to enable the database),
then we would have many events in the event_sourcing_light table that have not been processed -
possibly over days or weeks. We must not apply events that are that outdated!

So, if there is no cutoff yet, we write the current transformer to the cutoff,
so it doesn't get processed.

Ref: SRX-KNBOC7