fix(db): Do not run deployment migrations
Note that this migration is not technically required. It was just there to provide additional context to the user.
It only influences the UI, not what was actually deployed.

For now, we remove it. The migration itself seems fine, but the dection IF the migration is necessary is flawed,
leading the cd-service to potentially clone the git repo, even if it is not necessary.

Ref: SRX-YNEAHU