fix(db): Do not run deployment migrations (#2002)
Note that this migration is not technically required. It was just there
to provide additional context to the user. It only influences the UI,
not what was actually deployed.

For now, we remove it. The migration itself seems fine, but the detection
IF the migration is necessary is flawed, leading the cd-service to
potentially clone the git repo, even when it is not necessary. And that
is a very costly operation with large git repos.

Ref: SRX-YNEAHU