fix(db): fixed function InsertAppFun function override & cleaned up update of lightweight  apps on overview (#2066)
InserAppFun was only getting overridden when we performed custom
migrations. This led to a bug where when we initiated a kuberpult
instance without performing migrations (i.e. normal dev-env deployment
or make kuberpult locally), we would add an InsertAppFun that did not
update the overview.

Also added sql migration for deleting the old overview cache.

Ref: SRX-ZYK5HN