fix: move custom migrations to manifest-export
By "custom migrations" we mean those db migrations that can not be solved by plain SQL,
so they are implemented in go.

This provides the basic setup to move the custom migrations to the manifest-export-service.
* It creates a new endpoint to accept migrations, however the endpoint is not called yet.
* It provides the setup like reading/writing to the custom_migration_cutoff table

However, this change does not include actually moving the migrations from cd- to export-service.
This will be done in a separate PR.

Ref: SRX-V6RVYF