fix: only send metrics as background job (#2108)
Generating metrics for all apps and all envs can take quite some time.

We previously had some endpoints that directly send the metrics, in
order to have the metrics faster. Given how slow the metrics
calculations currently are, this is not feasible. We will now send
metrics only on the background task "RegularlySendDatadogMetrics".

Ref: SRX-Z8BHWD