fix(db): test complex environment queries (#2127)
There were multiple complex queries without unit tests. This PR adds
tests to 3 complex queries related to environments.

This PR added a support testing function for testing the
`TestReadReleasesWithoutEnvironments ` method.
This method is used for database migrations in case a release is in the
database without environments, but with any manifest. Our
`DBInsertRelease` method automatically adds any environment that has a
manifest to the release's environments list. The new method bypasses the
insertion for testing purposes.

Ref: SRX-THO0EO