fix(cd-service): improve get app details deployments (#2164)
The `GetAppDetails` endpoint collects information about the deployments
of an app.
One of these is the latest queued version. Currently the endpoint would
evaluate one deployment at a time and, when it was decided to be a valid
deployment, the database would be queried for this deployment's latest
queued version.

This PR queries the database for all latest deployment attempts. Then
when analyzing each deployment it just pulls from this information
cached in a map.

This might slow down GetAppDetails requests that have no current
deployments (since before we would not query the queued versions for
invalid deployments). However, if the app has some deployments, this
will translate in improved performance.

Ref: SRX-016Y9E