fix(cd): migration to delete foreign key on deployments and deployment history (#2239)
Parallel release trains spawn new transactions. This means that for a
release train we will have one outer transaction and N inner
transactions (based on the number of parallel processing go routines and
number of environments to process). As the outer transaction has not yet
been committed, inner transactions will not be able to view the most
recent event written to the event_sourcing_light table. This ALWAYS
results in a FK violation if anything is deployed by the release train,
as the transformer ID that the deployments try to reference does not
exist on the current database view of the inner transaction.

Rolling back the outer transaction causes inner transactions to also be
rolled back, so simply removing these fk_constrains works and kuberpult
won't be left in an inconsistent state (i.e. deployments exist that
don't reference any transformer on the event_sourcing_light table).

In my opinion, removing the FK constraint is the only solution, as we
want to keep using parallel transactions to perform the release trains
for different environments in parallel.

Ref: SRX-AC1K9Y