fix(manifest-export): deployments don't check for locks anymore on export service (#2284)
Fixed bug where the export service would crash with minimized git data
was on and we undeployed an application

We don't check for locks because either the Deployment:

* Was generated by a release train, in which case the release train
transformer checks what was deployed by the releases train,
* Was generated by a CreateApplicationVersion, that also checks what
deployments were generated by that transformer.
* Was generated by a CreateUndeployApplicationVersion, that also checks
what deployments were generated by that transformer.
* Was generated by the UI, which always has  api.LockBehavior = IGNORE.

This way we don't have to check for which transformer generated the
current we are processing, which would be
costly for big release trains that generate a lot of deployments.

Ref: SRX-HNENMA