fix(export-service): Update transformer eslversion upon failed esl event retry (#2295)
When the git sync status is enabled, we are able to retry events that
might have failed on the manifest export service.
When we retry an event, we pickup the failed event, copy it into the
event_sourcing_light table, but with a new eslVersion, so that the
manifest export service can naturally pick it up.
An issue arises when certain deployments depend on these failed events.
The way the manifest export service detects which deployments to write
to git is by filtering the deployments table by the deployments that the
transformer that is currently being processed has generated. If the
event is retried, a new eslVersion is given to the event and we lose the
reference to the deployments.
So if, for example, a release train transformer with tID: X fails and
gets retried, it gets given another tID: Y, but the deployments do not
get updated. As such, when we process the retried event, it queries the
database: "which deployments were generated by transformer with ID: Y",
and we get the answer: None. As such, the transformer successfully
executes and the apps appear synced, when in reality nothing was
deployed/pushed to the git repo

Ref: SRX-8FMYI8