feat(db): new db lock type and removed read-before-write
1) Replaced in-memory locks with db locks

Note that this feature is considered experimental.

With this change, we have three options in the helm chart (cd.lockType) to lock:
* "go": with an in-memory (golang) lock (default), initially implemented in #2301
* "db": with a postgres advisory lock
* "none": without any lock

The first option already existed, the other two options are new.
The postgres locks are an improvement over the go-locks, because they are not limited to the current pod/process.
Apart from that they work the same.

Both lock types ensure that only one destructive action (undeploy, delete env from app, delete env)
is running at a time.

2) Removed "Read-before-Writes"

There were 2 points where the cd-service read from the DB before writing:
When adding an app to an environment, and when removing an app from an environment.
Both have been simplified to one sql statement.

Ref: SRX-8JRR7Q