feat(db): replaced in-memory locks with db locks
Note that this feature is considered experimental.

With this change, we have three options in the helm chart (cd.lockType) to lock:
* "go": with an in-memory (golang) lock (default), initially implemented in #2301
* "db": with a postgres advisory lock
* "none": without any lock

The first option already existed, the other two options are new.
The postgres locks are an improvement over the go-locks, because they are not limited to the current pod/process.
Apart from that they work the same.

Both lock types ensure that only one destructive action (undeploy, delete env from app, delete env)
is running at a time.

Ref: SRX-8JRR7Q